[JB][CHANGED] Moved DiagnoseRequest screens inside screens folder and navigator inside navigators folder
