+++ src/screens/MainApp/DiagnoseRequest/screens/FinishRequest/index.js
