+++ src/screens/MainApp/DiagnoseRequest/screens/FinishRequest/styles.js
