+++ src/screens/MainApp/DiagnoseRequest/screens/ImageReview/Carousel/CarouselImage/index.js
