+++ src/screens/MainApp/DiagnoseRequest/screens/ImageReview/Carousel/CarouselImage/styles.js
