+++ src/screens/MainApp/DiagnoseRequest/screens/ImageReview/Carousel/index.js
