+++ src/screens/MainApp/DiagnoseRequest/screens/ImageReview/Carousel/styles.js
