+++ src/screens/MainApp/DiagnoseRequest/screens/ImageReview/ConfirmButton/index.js
