+++ src/screens/MainApp/DiagnoseRequest/screens/ImageReview/RenderedImage/index.js
