+++ src/screens/MainApp/DiagnoseRequest/screens/ImageReview/constants.js
