+++ src/screens/MainApp/DiagnoseRequest/screens/MainScreen/index.js
