+++ src/screens/MainApp/DiagnoseRequest/screens/MainScreen/styles.js
