+++ src/screens/MainApp/DiagnoseRequest/screens/NoPhotoDisclaimer/styles.js
